refactor: global/common -> global/domain 변경
