+++ src/main/java/com/ururulab/ururu/global/domain/entity/enumerated/EnumParser.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.global.common.entity.enumerated;
+package com.ururulab.ururu.global.domain.entity.enumerated;
 
 import java.util.Arrays;
