+++ src/main/java/com/ururulab/ururu/global/domain/entity/enumerated/Gender.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.global.common.entity.enumerated;
+package com.ururulab.ururu.global.domain.entity.enumerated;
 
 public enum Gender {
 	MALE,
