+++ src/main/java/com/ururulab/ururu/global/domain/repository/TagRepository.java
@@ -1,8 +1,8 @@
-package com.ururulab.ururu.global.common.repository;
+package com.ururulab.ururu.global.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ururulab.ururu.global.common.entity.Tag;
+import com.ururulab.ururu.global.domain.entity.Tag;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
 }
