+++ src/main/java/com/ururulab/ururu/member/domain/entity/MemberPreference.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.member.domain.entity;
 
-import com.ururulab.ururu.global.common.entity.BaseEntity;
+import com.ururulab.ururu.global.domain.entity.BaseEntity;
 import com.ururulab.ururu.member.domain.entity.enumerated.PurchaseFrequency;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Max;
