+++ src/main/java/com/ururulab/ururu/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.ururulab.ururu.member.service;
 
 import com.ururulab.ururu.auth.dto.info.SocialMemberInfo;
-import com.ururulab.ururu.global.common.entity.enumerated.Gender;
+import com.ururulab.ururu.global.domain.entity.enumerated.Gender;
 import com.ururulab.ururu.member.domain.dto.request.MemberRequest;
 import com.ururulab.ururu.member.domain.dto.response.*;
 import com.ururulab.ururu.member.domain.entity.Member;
