+++ src/main/java/com/ururulab/ururu/order/domain/entity/Cart.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.order.domain.entity;
 
-import com.ururulab.ururu.global.common.entity.BaseEntity;
+import com.ururulab.ururu.global.domain.entity.BaseEntity;
 import com.ururulab.ururu.member.domain.entity.Member;
 import com.ururulab.ururu.order.domain.policy.CartPolicy;
 import jakarta.persistence.*;
