+++ src/main/java/com/ururulab/ururu/order/domain/entity/OrderHistory.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.order.domain.entity;
 
-import com.ururulab.ururu.global.common.entity.BaseEntity;
+import com.ururulab.ururu.global.domain.entity.BaseEntity;
 import com.ururulab.ururu.order.domain.entity.enumerated.OrderStatus;
 import com.ururulab.ururu.order.domain.policy.OrderHistoryPolicy;
 import jakarta.persistence.*;
