+++ src/main/java/com/ururulab/ururu/payment/domain/entity/enumerated/PointSource.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.payment.domain.entity.enumerated;
 
-import com.ururulab.ururu.global.common.entity.enumerated.EnumParser;
+import com.ururulab.ururu.global.domain.entity.enumerated.EnumParser;
 
 public enum PointSource {
     GROUPBUY,   // 공동구매
