+++ src/main/java/com/ururulab/ururu/product/domain/entity/ProductCategory.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.product.domain.entity;
 
-import com.ururulab.ururu.global.common.entity.BaseEntity;
+import com.ururulab.ururu.global.domain.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
