+++ src/main/java/com/ururulab/ururu/review/controller/ReviewController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.ururulab.ururu.global.common.entity.enumerated.Gender;
+import com.ururulab.ururu.global.domain.entity.enumerated.Gender;
 import com.ururulab.ururu.member.domain.entity.Member;
 import com.ururulab.ururu.member.domain.entity.enumerated.Role;
 import com.ururulab.ururu.member.domain.entity.enumerated.SocialProvider;
