+++ src/main/java/com/ururulab/ururu/review/domain/entity/ReviewImage.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.review.domain.entity;
 
-import com.ururulab.ururu.global.common.entity.BaseEntity;
+import com.ururulab.ururu.global.domain.entity.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
