+++ src/main/java/com/ururulab/ururu/review/domain/entity/enumerated/AgeGroup.java
@@ -1,6 +1,6 @@
 package com.ururulab.ururu.review.domain.entity.enumerated;
 
-import com.ururulab.ururu.global.common.entity.enumerated.EnumParser;
+import com.ururulab.ururu.global.domain.entity.enumerated.EnumParser;
 
 public enum AgeGroup {
 	TEENS, TWENTIES, THIRTIES, FORTIES_PLUS;
