+++ src/main/java/com/ururulab/ururu/review/service/ReviewService.java
@@ -2,7 +2,6 @@ package com.ururulab.ururu.review.service;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.context.ApplicationEventPublisher;
@@ -10,10 +9,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.ururulab.ururu.global.common.entity.Tag;
-import com.ururulab.ururu.global.common.entity.enumerated.Gender;
-import com.ururulab.ururu.global.common.entity.enumerated.SkinType;
-import com.ururulab.ururu.global.common.repository.TagRepository;
+import com.ururulab.ururu.global.domain.entity.Tag;
+import com.ururulab.ururu.global.domain.entity.enumerated.Gender;
+import com.ururulab.ururu.global.domain.entity.enumerated.SkinType;
+import com.ururulab.ururu.global.domain.repository.TagRepository;
 import com.ururulab.ururu.member.domain.entity.Member;
 import com.ururulab.ururu.member.domain.entity.enumerated.Role;
 import com.ururulab.ururu.member.domain.entity.enumerated.SocialProvider;
@@ -27,7 +26,6 @@ import com.ururulab.ururu.review.domain.entity.enumerated.AgeGroup;
 import com.ururulab.ururu.review.domain.repository.ReviewRepository;
 import com.ururulab.ururu.review.event.ReviewCreatedEvent;
 
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 
 @Service
